refactor(desktop): extract startup presentation policies
Move the voice-context flight owner and Home stage transitions out of oversized controllers, ratcheting both line-count baselines downward without changing behavior.
