+++ desktop/macos/Desktop/Sources/MainWindow/Pages/DashboardPage.swift
@@ -2234,101 +2234,6 @@ private enum HomeDestinationProminence {
   case quiet
 }
 
-enum HomeStageMode: Equatable {
-  case hub
-  case chat
-  case connect
-
-  /// Whether the user-facing collapse catchers (click-outside + Esc) mount.
-  /// Only a panel that can collapse to a *different* resting surface gets a
-  /// catcher. The hub is the base surface, never an overlay: mounting a
-  /// catcher over hub-with-history would invert the gesture and make a stray
-  /// click or Esc *open* the chat.
-  static func collapseCatcherActive(mode: HomeStageMode, resting: HomeStageMode) -> Bool {
-    mode != resting && mode != .hub
-  }
-
-  var automationLabel: String {
-    switch self {
-    case .hub: return "hub"
-    case .chat: return "chat"
-    case .connect: return "connect"
-    }
-  }
-}
-
-enum HomeHistoryPresentationPolicy {
-  static func restingMode(isLoading: Bool, messageCount: Int) -> HomeStageMode {
-    !isLoading && messageCount > 0 ? .chat : .hub
-  }
-}
-
-/// Shared "drop from the top" motion for stage panels: a short slide with a
-/// slight top-anchored scale and fade — deliberate, not a full-height fly-in.
-private struct HomeStageDropModifier: ViewModifier {
-  let offsetY: CGFloat
-  let scale: CGFloat
-  let opacity: Double
-
-  func body(content: Content) -> some View {
-    content
-      .offset(y: offsetY)
-      .scaleEffect(scale, anchor: .top)
-      .opacity(opacity)
-  }
-}
-
-extension AnyTransition {
-  fileprivate static var homeDropFromTop: AnyTransition {
-    .modifier(
-      active: HomeStageDropModifier(offsetY: -46, scale: 0.97, opacity: 0),
-      identity: HomeStageDropModifier(offsetY: 0, scale: 1, opacity: 1)
-    )
-  }
-
-  fileprivate static var homeHubFade: AnyTransition {
-    .modifier(
-      active: HomeStageDropModifier(offsetY: 14, scale: 1, opacity: 0),
-      identity: HomeStageDropModifier(offsetY: 0, scale: 1, opacity: 1)
-    )
-  }
-
-  /// During initial history restoration, the useful hub leaves upward while
-  /// the completed conversation rises from below into the same stage.
-  fileprivate static var homeHubStage: AnyTransition {
-    .asymmetric(
-      insertion: .modifier(
-        active: HomeStageDropModifier(offsetY: 14, scale: 1, opacity: 0),
-        identity: HomeStageDropModifier(offsetY: 0, scale: 1, opacity: 1)
-      ),
-      removal: .modifier(
-        active: HomeStageDropModifier(offsetY: -54, scale: 0.98, opacity: 0),
-        identity: HomeStageDropModifier(offsetY: 0, scale: 1, opacity: 1)
-      )
-    )
-  }
-
-  fileprivate static var homeChatRise: AnyTransition {
-    .asymmetric(
-      insertion: .modifier(
-        active: HomeStageDropModifier(offsetY: 54, scale: 0.98, opacity: 0),
-        identity: HomeStageDropModifier(offsetY: 0, scale: 1, opacity: 1)
-      ),
-      removal: .modifier(
-        active: HomeStageDropModifier(offsetY: -28, scale: 0.98, opacity: 0),
-        identity: HomeStageDropModifier(offsetY: 0, scale: 1, opacity: 1)
-      )
-    )
-  }
-
-  fileprivate static var homeSuggestionsFade: AnyTransition {
-    .modifier(
-      active: HomeStageDropModifier(offsetY: 10, scale: 1, opacity: 0),
-      identity: HomeStageDropModifier(offsetY: 0, scale: 1, opacity: 1)
-    )
-  }
-}
-
 /// The persistent home ask bar: a pill-shaped chat input with attachments
 /// (paperclip + drag-drop, same limits as the chat page), a send/stop action,
 /// and the Connect toggle living inside the pill.
